# web/backend/static/openapi.yaml
openapi: 3.1.0
info:
  title: ESA Host Brief Viewer
  version: 0.2.0
  description: 'ESA Incentive Pipeline backend API.


    **Authentication**


    - Host endpoints: require `esa_session` cookie (set via `POST /api/login`). Valid
    24h.

    - Admin endpoints: require `esa_admin_session` cookie (set via `POST /api/admin/login`).
    Valid 1h.


    **Base URL**: `https://esa-hosting.codescales.xyz`


    **AI/Tool discovery**


    - OpenAPI spec (this file): `GET /openapi.yaml`

    - Natural-language context doc: `GET /llms.txt`

    - Interactive UI: `GET /docs`'
  contact:
    name: ESA Incentive Pipeline
    url: https://github.com/BongoEADGC6/esa-incentive-pipeline
paths:
...
  /api/incentives:
    get:
      tags:
      - incentives
      summary: List Incentives
      operationId: list_incentives_api_incentives_get
      parameters:
      - name: run_slug
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Run Slug
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Status
      - name: category
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Category
      - name: stream
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Stream
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncentiveDTO'
                title: Response List Incentives Api Incentives Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - incentives
      summary: Create Incentive
      operationId: create_incentive_api_incentives_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncentiveCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncentiveDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/incentives/{uuid}:
    get:
      tags:
      - incentives
      summary: Get Incentive
      operationId: get_incentive_api_incentives__uuid__get
      parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          title: Uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncentiveDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - incentives
      summary: Patch Incentive
      operationId: patch_incentive_api_incentives__uuid__patch
      parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          title: Uuid
      - name: esa_admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Esa Admin Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncentivePatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncentiveDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - incentives
      summary: Delete Incentive
      operationId: delete_incentive_api_incentives__uuid__delete
      parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          title: Uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncentiveDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/briefs/{slug}:
    get:
      tags:
      - briefs
      summary: Get Brief
      operationId: get_brief_api_briefs__slug__get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/briefs:
    get:
      tags:
      - briefs
      summary: List Briefs
      operationId: list_briefs_api_briefs_get
      parameters:
      - name: shift
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Shift
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefIndexResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IncentiveCreateRequest:
      properties:
        run_slug:
          type: string
          title: Run Slug
        incentive_text:
          type: string
          title: Incentive Text
        incentive_category:
          type: string
          title: Incentive Category
          default: ''
        valid_for_game:
          type: string
          title: Valid For Game
          default: ''
        incentive_estimate:
          type: string
          title: Incentive Estimate
          default: ''
        status:
          type: string
          title: Status
          default: ''
      type: object
      required:
      - run_slug
      - incentive_text
      title: IncentiveCreateRequest
      description: Fields required to create a new incentive.
    IncentiveDTO:
      properties:
        scheduled:
          type: string
          format: date-time
          title: Scheduled
        game:
          type: string
          title: Game
        category:
          type: string
          title: Category
        stream:
          type: string
          title: Stream
        runner_display:
          type: string
          title: Runner Display
        runner_twitch:
          type: string
          title: Runner Twitch
        runner_discord:
          type: string
          title: Runner Discord
        runner_slug:
          type: string
          title: Runner Slug
          default: ''
        incentive_text:
          type: string
          title: Incentive Text
        incentive_category:
          type: string
          title: Incentive Category
        valid_for_game:
          type: string
          title: Valid For Game
        incentive_estimate:
          type: string
          title: Incentive Estimate
        needs_approval:
          type: string
          title: Needs Approval
        status:
          type: string
          title: Status
        submission_id:
          type: string
          title: Submission Id
        uuid:
          type: string
          title: Uuid
        run_slug:
          type: string
          title: Run Slug
          default: ''
        participants:
          items:
            $ref: '#/components/schemas/ParticipantDTO'
          type: array
          title: Participants
          default: []
      type: object
      required:
      - scheduled
      - game
      - category
      - stream
      - runner_display
      - runner_twitch
      - runner_discord
      - incentive_text
      - incentive_category
      - valid_for_game
      - incentive_estimate
      - needs_approval
      - status
      - submission_id
      - uuid
      title: IncentiveDTO
    IncentivePatch:
      properties:
        incentive_category:
          anyOf:
          - type: string
          - type: 'null'
          title: Incentive Category
        valid_for_game:
          anyOf:
          - type: string
          - type: 'null'
          title: Valid For Game
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        incentive_estimate:
          anyOf:
          - type: string
          - type: 'null'
          title: Incentive Estimate
      type: object
      title: IncentivePatch
      description: Editable fields on an incentive. All optional — only set fields
        are updated.
...
    ParticipantDTO:
      properties:
        slug:
          type: string
          title: Slug
          default: ''
        display_name:
          type: string
          title: Display Name
          default: ''
        twitch:
          type: string
          title: Twitch
          default: ''
        discord:
          type: string
          title: Discord
          default: ''
        twitter:
          type: string
          title: Twitter
          default: ''
        pronouns:
          type: string
          title: Pronouns
          default: ''
        pronunciation:
          type: string
          title: Pronunciation
          default: ''
        submission_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Submission Id
        match_confidence:
          type: string
          title: Match Confidence
          default: ''
      type: object
      title: ParticipantDTO
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    hostSession:
      type: apiKey
      in: cookie
      name: esa_session
      description: Set via POST /api/login. Valid 24h.
    adminSession:
      type: apiKey
      in: cookie
      name: esa_admin_session
      description: Set via POST /api/admin/login. Valid 1h.
servers:
- url: https://esa-hosting.codescales.xyz
  description: Production
...